requirements.txt
pymongo[srv]
discord.py
flask
asyncpraw
dnspython